# Pre-allocate res change-items based on statistic